refactor: move ANSI to internal
- Moved all ANSI code constants to internal package
- Moved ANSI styles (bold, italic, etc) from `colors/` to `styles/`
