+++ color/background/purple_test.go
@@ -3,8 +3,8 @@ package background
 import (
 	"testing"
 
-	ansi "github.com/sitnikovik/paints/ansi/color"
-	ansiBackground "github.com/sitnikovik/paints/ansi/color/background"
+	ansi "github.com/sitnikovik/paints/internal/ansi/color"
+	ansiBackground "github.com/sitnikovik/paints/internal/ansi/color/background"
 )
 
 func TestPurple(t *testing.T) {
